Allow the force_encoding plugin to work when refreshing
Before, the force_encoding plugin only worked by overriding load and
typecast_value, so it didn't handle the case where refresh was
called, either explicitly or implicitly inside save.

Add a Model#set_values private method that sets @values to the
given hash, and have load and refresh both use it.  Modify the
force_encoding plugin to override it.  Since it no longer overrides
load, it wouldn't work correctly with the identity map/lazy
attributes plugins, so override a method that the identity map
plugin uses so that the force encoding plugin will work with it.
You need to load the force encoding plugin after the identity
map plugin for this to work.