Revert the MSSQL JOIN USING emulation, replace with standard support
Standard support for emulating JOIN USING is better since multiple
database's don't support JOIN USING (MSSQL and H2).  Note that
emulating JOIN USING really sucks, for 2 reasons:

1) It doesn't combine the output columns, so you must qualify any
column references to the JOIN USING columns.  This makes using this
support with the class_table_inheritance possible but still
problematic.

2) The support assumes the last joined table has the column, when it
could be any table previously joined.  That's a poor assumption, but
the only way to work around it is to wrap the current dataset in a
subselect, which could break queries that use qualified identifiers
with previously joined tables.