Add more checks to the add_ and remove_ association methods
The add_ and remove_ methods both now check if the passed object is
of the correct class.

The remove_ method now checks that the passed object is already
associated.  It's probably broken for many_to_many associations, but
that will be fixed soon.

Also, you can now pass an array to the remove_ method to handle
removing an associated object which uses a model with composite
primary keys.

While here, update the exception messages to show specific instance
information.