Fix remove_ method for many_to_many associations, other minor fixes
The remove_ method broke for many many_to_many associations with the
change to check that the passed object was already associated, both
in the case where a primary key is passed and the case when an
existing object is passed.

many_to_many associations don't need this strictness check, because
attempting to remove the association for a passed object that
isn't associated will have no negative effect.  It will attempt
to remove rows from the join table, which won't exist.

To do this, add a remove_should_check_existing? method to the
AssociationReflection classes.

While here, add specs for the many_to_many case, and add integration
tests.  Also, fix the nested attributes specs which broke in a couple
places.