Allow easy overriding of the default error messages in the validation_helpers plugin
Previously, the validation_helpers plugin used hard coded strings for
error messages.  Now, it stores the messages as procs in the
MESSAGE_PROCS plugin module level constant hash (keyed on validation
type symbol).  The values of the hash are lambdas that may or may not
take an argument (depending on validation type), and return the
validation error message string to use.  Here's an example of
overriding the default error message for validates_presence:

  Sequel::Plugins::ValidationHelpers::MESSAGE_PROCS[:presence] = \
    lambda{"was not entered"}

You can also modify the error messages on a per model basis by
overriding the default_validation_error_message_proc private
instance method.  See the source code if you want to do this.

Additionally, you can now specify the :message option as a proc
that operates just like a default error message proc.  The main
use case for that is if you want to reuse the proc in many different
calls with different arguments.

I choose to use procs instead of format strings as that gives you
the maximum amount of flexibility (at the expense of performance).