Fix marshalling of model objects so that they load correctly, but you need to call marshallable! first
The previous fix of not raising an error when marshalling the model
instance's dataset didn't work as it broke loading.  Instead of
writing custom _dump and _load methods, add a marshallable! instance
method that just sets the model instance's dataset to nil.  Leave the
changes to Model::Errors, since they aren't a bad way to handle
things.

Make the specs test loading the marshalled data, so we can make sure
things work.