Don't allow modifying keys related to the association when updating existing associated objects in the nested_attributes plugin
This fixes a possible security issue in the nested attributes plugin,
allowing you to modify an associated object's fields related to the
association, which could allow you to disassociate it from the
current record or associate it to a different record.

This security issue is only likely to be present in one_to_many
assocations, but it can happen for other associations if you use
non default primary key fields in the association.  Here's an
example of code that exploits the vulnerability:

  Artist.one_to_many :albums
  Artist.plugin :nested_attributes
  Artist.nested_attributes :albums
  artist = Artist.create
  artist2 = Artist.create
  album = Album.create
  artist.albums_attributes = [{:id=>album.id, :artist_id=>artist2.id}]
  artist.save