Add support for a :fields option in the nested attributes plugin, and only allow updating of the fields specified
I'm a big believer in using set_only for all mass assignment,
specifying explicitly the fields allowed in that mass assignment
call.  This patch allows you to do that when using the
nested_attributes, by specifying a :fields option which should
be an array of symbols, which changes the internals to use
set_only with the fields specified, instead of set.

This fixes the spec for the previous commit, which only passed by
accident.  It also reorganizes some of the private
nested_attributes methods.