Make add_ association method not add the associated object to the cached array if it's already there
This makes the most sense for one_to_many associations, since those cannot
be associated more than once.  For many_to_many associations, I think it
is a better default, as it's a rare case that you want duplicate items
in the association.  If you do want duplicate items for a many_to_many
association and you want it to work well with caching, please bring it
up on the Sequel mailing list and we'll work it out.

Note that this exact thing was done for reciprocal associations, so
Sequel is more consistent with this change.