Allow easier overriding of default options used in the validation_helpers plugin
This breaks backwards compatibility if you were relying on the
change made on October 20 to set the default validation error
messages.  This generalizes support for overriding any of the
default options.  As before, you can override it on a global level:

  Sequel::Plugins::ValidationHelpers::DEFAULT_OPTIONS[:format].
    merge!(:message=>"incorrect format", :allow_missing=>true)

You can also override it on a per class basis via:

  class Album < Sequel::Model
    plugin :validation_helpers
    DEFAULT_VALIDATION_OPTIONS = {
      :format=>{:message=>"incorrect format", :allow_missing=>true}}

    private

    def default_validation_helpers_options(type)
      super.merge(DEFAULT_VALIDATION_OPTIONS[type] || {})
    end
  end

This removes the MESSAGE_PROCS constant from the ValidationHelpers
module, and removes the private instance method
default_validation_error_message_proc.  It changes the API for
validation_error_message and adds the
default_validation_helpers_options and
validatable_attributes_for_type private instance methods.

I chose not to change the validatable_attributes API, even though
it could be simplified now, because while the method is private,
I expect there may be consumers other than the default validations.