Use save_changes instead of save when updating existing associated objects in the nested_attributes plugin
This should be significantly faster for the case where you are
displaying a complex form to the user, and they make no changes to
the existing associated records when they submit it.

This shouldn't have a negative effect on production code, unless
save_changes doesn't work correctly for the associated objects.
It may break test code that checks for specific SQL or columns in
the update statement.