Make Dataset#group_and_count handle arguments other than Symbols
I think with a past change to handle implicit aliases in Symbol
arguments, it no longer worked with other objects, as it assumed
all arguments were Symbols.  That is not a valid assumption.

Introduce Dataset#unaliased_identifier private method that
handles both implicit aliases in symbols and AliasedExpression
objects, returning the unaliased argument, and have group_and_count
use it.