Add NULL, NOTNULL, TRUE, SQLTRUE, FALSE, and SQLFALSE constants
The main advantage here for most people is NOTNULL.  This allows the
following to work:

  include Sequel::SQL::Constants
  DB[:table].where(:a=>'1', :b=>NOTNULL)

Before, the shortest way to do this was:

  DB[:table].where(:a=>'1').exclude(:b=>nil)

It may make the code more descriptive:

  DB[:table].where(:b=>NULL)
  # compared to
  DB[:table].where(:b=>nil)

This gives the option to use SQL terminology instead of ruby
terminology.

The other advantage of using the constants it that they handle
operators and methods like other Sequel::SQL objects:

  NULL & SQLFALSE # BooleanExpression => "(NULL AND FALSE)"
  nil & false     # false

  NULL + :a       # NumericExpression => "(NULL + a)"
  nil + :a        # raises NoMethodError
  NULL.sql_string + :a  # StringExpression => "(NULL || a)"
  NULL.as(:b)     # AliasedExpression => "NULL AS b"

For complex systems that want to represent SQL boolean objects
in ruby (where you don't know exactly how they'll be used), using
the constants is recommended.

In order not to be too verbose, including Sequel::SQL::Constants
is recommended.  It's not done by default, but you can still
reference the constants under the main Sequel module by default
(e.g. Sequel::NULL).