No longer use native autoreconnection in the mysql adapter
Native autoreconnection has problems with prepared statements, where
a new native connection is used behind Sequel's back, so Sequel
thinks the prepared statement has already been defined on the
connection, when it fact it hasn't.  Any other changes that affect
the state of the connection will be lost when native autoreconnection
is used as well.

Sequel's connection pool already handles reconnection if it detects a
disconnection.  This commit also adds an additional exception message
to recognize as a disconnect.