Fix deleting from joined datasets on MySQL
While here, no longer include joins in delete_from_sql by default.
Since all of the adapters that support modifying joins override it,
there's no reason to include it.

Also, remove the order on the joined dataset in the integration
test so that updating works on MySQL (which allows order clauses
in update statements, but not when multiple tables are used).

Note that MySQL allows deleting from multiple tables at once, but
Sequel doesn't support this.  Sequel assumes the first table in
FROM is the one being deleted from, and the rest are just for
selecting rows.  This is for consistency with how Sequel works on
PostgreSQL and MSSQL.