Add Dataset#sequence to the shared Oracle Adapter, for returning autogenerated primary key values on insert (Fixes #280)
Oracle is like PostgreSQL in that autogenerated primary key values
are handled by sequences.  Unlike Sequel's shared PostgreSQL adapter,
Sequel's shared Oracle adapter doesn't handle automatically
determining the primary key value based on the table.  That may be
possible to fix, but without access to an Oracle database it would
be difficult.

Dataset#sequence provides a workaround for the problem, allowing the
manual specification of the sequence. Dataset#insert is then
overridden to check for an associated sequence, and return the most
recently inserted value upon insert.