Handle multiple statements in a single query in the native MySQL adapter in all cases, not just when selecting via Dataset#each
Before, if you used Database#run or another method to run queries,
using multiple statements would cause a commands out of sync error.
This contains changes to always check for multiple statement result
sets when executing queries.

Note that the native MySQL adapter is one of the few (maybe the
only) adapters that supports sending multiple statements in a single
query, and using the feature is not encouraged.