Allow removing servers/shards from a database and connection pool, and schedule disconnection of allocated connections
There are two main changes in this commit.  The first is the
scheduling of disconnection for currently allocated connections.
Previously, Database#disconnect would only disconnect available
connections, with no affect on currently allocated connections.
This commit changes that so that disconnect schedules currently
allocated connections to be disconnected as soon as they are
returned to the connection pool.

The second part is remove_servers, which removes existing shards
from the connection pool and database.  This is basically the same
as disconnecting, except that future attempts to acccess the server
will use the :default server instead.

Another small change is that #disconnect now accepts an option
hash and respects a :server option that can be a symbol or array
of symbols specifying server/shard names.  If used, only those
servers/shards will be disconnected.

The internals of the connection pool did change somewhat.  @servers
is now a hash which maps nonexistent server names to :default.
@allocated and @available_connections hashes no longer have default
procs.  Instead, when pool.hold is called, @servers is used to
transform the server name before anything else is done.

A consequence of this is that allocated and available_connections
return nil instead of an array or hash if they are called with a
nonexistent server.  This could be changed back, but I think the
current behavior is better.  created_count/size still does the
translation, though.

ConnectionPool#sync private method is added to cut down on some
code duplication.  ConnectionPool#remove no longer uses the mutex
internally, so callers need to have the mutex before calling it.
ConnectionPool#release now takes 3 arguments instead of 2.

While here, beef up the documentation for some of the private
methods.