Catch cases in the postgres adapter where exceptions weren't converted or raised appropriately
This adds check_disconnect_errors to the Postgres::Adapter class,
and check_database_errors to the Postgres::Database class.  The
former rescues PGErrors and converts them to
DatabaseDisconnectErrors if the connection status cannot be
determined or is not OK.  The latter converts exceptions to
regular DatabaseErrors.  They both yield to their blocks.
This cuts down on the amount of manually rescuing in the adapter.

Specifically, this patch fixes two problems:

1) PGErrors that weren't converted when preparing a statement or
executing a prepared statement.

2) A DatabaseDisconnectError not being raised for a non-OK status.