Don't use the OUTPUT clause on SQL Server versions that don't support it (pre-2005) (Fixes #281)
The OUTPUT clause was only used implicitly in a single case,
Dataset#insert_select, so otherwise OUTPUT would not be used unless
Dataset#output was called manually.  Unfortunately, model object
creation uses insert_select, so creating a model object on SQL
Server 2000 failed.

This commit checks the server version to make sure OUTPUT is
supported before using it.  If Dataset#output is called explicitly
when not supported, an Error is raised.