Fix performance issue for prepared statements
Prepared statements are supposed to cache their SQL for performance
reasons, but in the general case, that caching was being handled in
copies of the prepared statement that were created when calling the
statement, instead of in the prepared statement object itself. This
probably happened when Dataset#bind was added a few versions back,
but it's possible it's been around for longer.

Thanks to Susumu Iima for the heads up.

This requires a small spec change, and any custom adapters that
implement native prepared statements will need to make sure to call
prepared_sql on the prepared statement object before returning it
in Dataset#prepare.