Fix RDoc for class level accessors
Previously, in a few places in the Sequel where class-level accessor
methods are created using attr_* on the singleton class, class-level
RDoc was used instead of method-level RDoc.  I recall this was
because previously there was a metaprogramming method that added
the class-level accessors (which RDoc ignored), so class-level RDoc
was the only way to document it.  Now that standard ruby methods that
RDoc understands are used, we can switch to method-level RDoc.

While here, mention the :charset, :collate, and :engine options that
Database#create_table respects when using MySQL, in the RDoc for
setting the defaults for these values.