Major changes: connection pool refactoring, thread safe requiring
The main change this commit makes is refactoring
Sequel::ConnectionPool from being a single class to being a class
hierarchy with 4 subclasses, for each combination of thread usage
and sharding usage.

Before, Sequel::SingleThreadedPool existed as a single-threaded pool,
but both it and the standard connection pool implemented sharding.
Unfortunately, supporting sharding hurts performance, and very
few people use it.  The main purpose of this commit is to make
the default connection pool unsharded, so most users don't
have to pay a performance penalty for functionality they are not
using.  The sharded connection pools are used automatically if
the :servers option is used when connecting to the database.

In addition to the lack of sharding in the default connection pool,
the new default connection pool also no longer contains the code
to schedule disconnections of in-use connections.  That code still
remains in the sharded connection pool, so if you need that
functionality, you should pass a :servers=>{} option when connecting
if you are not already sharding.

Because of the connection pool differences, if you plan on sharding
at any point, you must pass a :servers option when connecting to the
database (even if it is the empty hash).  You can no longer just
call add_servers afterward to add additional shards.

Another change made to all connection pools is that converting of
non-StandardError based exceptions to RuntimeErrors is no longer
done.  This was already turned off for all adapters I test with,
and I don't think it serves any purpose.

Yet another change made was the removal of the connection_proc and
disconnection proc accessors from ConnectionPool.  Now, a block
argument to ConnectionPool#initialize is required, and a
disconnection_proc must be specified via the :disconnection_proc
option to initialize.

Because of the connection pool changes, ConnectionPool.new should
no longer be utilized.  Instead ConnectionPool.get_pool should be
used, which will return an appropriate pool subclass instance.

You can now easily provide your own connection pool class, using
the :pool_class option when connecting to the database.

The Sequel::SingleThreadedPool constant is no longer defined.

This commit makes Sequel.extension, Sequel::Model.plugin, and
adapter and driver loading operate in a thread safe manner.  To do
so, it introduces a couple module methods to the Sequel module:

  ts_require - Thread safe version of Sequel.require
  tsk_require - Thread safe version of standard require

It's possible that these changes can cause errors if you have
a custom plugin that loads a Sequel.extension (and other
possible examples).  Because Sequel.extension and
Sequel::Model.plugin are both thread safe, but the mutex used is not
reentrant, plugins that require extensions need to use
Sequel.nts_extension.

This commit makes some small unrelated changes:

* Empty query parameter keys in a connection string are now ignored
  instead of raising an exception.

* The options hash passed to Sequel.connect is no longer modified.
  Also, there may be additional options present in Database#opts
  hash that weren't provided in the original hash passed to
  initialize (e.g. :disconnection_proc and :single_threaded).

* The default version of Database#connect now matches the API used
  by the adapters.

* Sequel::Model.plugin_gem_location and plugin_gem_location_old have
  been removed.

* A spec was added for raising ValidationFailed with a String,
  bringing line coverage back to 100%.