Make internal thread safe requiring methods reentrant
Under certain circumstances, the thread safe requiring can result
in a thread trying to grab the require mutex when it already has it.
This can happen in plugins that load extensions.  Previously,
Sequel.nts_extension was added to work around this, but with this
commit, it has been removed.  Instead, a private singleton method
check_requiring_thread has been added to the main Sequel module,
which ts_require and tsk_require use.

Design is similar to the connection pool and transaction reentrancy
(using return yield), except that only a single thread is allowed.