Handle IN/NOT IN with an empty array
Previously, Sequel may not have handled IN/NOT correctly when given
an empty array.  Also, it may not have handled IN/NOT correctly
for multiple columns and datasets when the database didn't directly
support multiple columns for IN/NOT IN.

This changes the code to handle the empty array as a special case,
using col != col for correct NULL semantics in the IN case, and 1 = 1
for the NOT IN case. The idea for this way of handling things came
from reading SQLAlchemy's documentation.

Sequel should now correctly work with all types of IN/NOT IN and
arrays or datasets. Even the case where the database doesn't support
multiple columns for IN/NOT IN, and a dataset is used instead of an
array is now handled correctly by sending the subquery separately,
getting the results as an array and then handling it like an array.