Allow limit and offset to be bound variables
Before, Sequel would call to_i on all limits and offsets.  This was
originally done for security so that if a programmer decided to
pass an untrusted string to Dataset#limit, it would not be used
verbatim.  Since bound variables in Sequel are specified with a
symbol starting with $, they didn't work correctly with
Dataset#limit.

This commit changes the code so that only regular strings are
converted to integers, with other objects being literalized normally.
This allows you to use a database function as a limit or offset if
the dataset supports it, as well as allows it to work with prepared
statements and bound variables.  Small changes were needed to the
MSSQL shared adapter, since offset support is emulated there.

This likely does not work on Oracle, but as I have no ability to test
changes to the Oracle adapter, I didn't attempt to modify it.