Combine multiple complex expressions for simpler SQL and object tree
I've thought about doing this for a long time, and it turned out to
be really easy.  Before, if you had an expressions like:

  :a + :b + :c

Sequel would literalize it to:

  ((a + b) + c)

because that's how ruby parses it.  With this commit, if you use one
of the n-arity operators (AND, OR, ||, +, -, *, or /), and you apply
that operator to a complex expression with the same operator, Sequel
will merge the arguments for the operators to reduce the nesting
level:

  (a + b + c)

This makes the SQL simpler and easier to read, and potentially
reduces the number of objects in complex expression trees.