Add Model.one_to_one option for creating model associations
This adds the one_to_one class method, which just adds a one_to_many
association with the :one_to_one option.  It also has some behavior
changes:

Instead of the :one_to_one option creating an add_ association
method, it just creates the setter method, which doesn't call add_.
Before the behavior was inconsistent in that it called the add_
method but not the remove_ method for removing existing options
(which is what the many_to_one setter does).  I'm going to be
changing the many_to_one and one_to_one associations to not use
add_ and remove_, but that's coming later.

The association_dataset method for one_to_one associations is
no longer private.

Using the many_to_one setter when the reciprocal association is a
one_to_one association with a different associated object already in
the reciprocal cached association no longer causes an error.

Instead of raising an error if multiple objects are returned by the
dataset, it defaults to selecting only 1 item in the dataset. If for
some reason the cached association has more than one record, it
returns the last record.  The reason it uses the last record instead
of the first record is to make it work correctly with the many_to_one
setter reciprocal handling.