Use before_set and after_set callbacks for many_to_one and one_to_one associations
Before, the many_to_one association setter called before and after
remove hooks for the currently associated object (sending a database
query to get the currently associated record if necessary), and
before and after add hooks for the current record.  This has been an
issue in the past, and it's generally not necessary.  The new before
and after set callbacks make the handling more like the how the
before and after add and remove callbacks are handled by the add and
remove methods.

This changes the internals to set_associated_object to simplify it
to use the new callbacks.  It also changes the one_to_one setter
to split it into public and private parts, adding a
set_one_to_one_associated_object method that is similar to
set_associated_object but different enough to require a standalone
method.

The private one_to_one setter has been improved somewhat in regards
to handling nil, but it's not good yet.  This will be fixed soon.
Another fix for the one_to_one setter is that it now sets other
associated records to nil before the current record, which could
possibly be better on databases that don't delay foreign key
constaint checking until the end of the transaction (or when
transactions aren't being used).