Make one_to_one a real association type
Previously, one_to_one was a one_to_many association with some
behavior changes.  This commit doesn't change much, but it does make
one_to_one it's own association.  Calling one_to_many with the
:one_to_one option will create a one_to_one association.

Make the association methods use the same API as associate, to keep
them consistent with each other.

This shouldn't be a significant external change, other than
one_to_one association type being :one_to_one instead of
:one_to_many.  One significant internal change is that
AssociationReflection#reciprocal_type is now allowed to return an
array of symbols, in which case any of those association types can
be a candidate to be a reciprocal.