Cache one_to_one associations using an object instead of an array
This is nearing the end of the one_to_one commits, but we still have
a couple to go.  Because one_to_one associations were originally
just one_to_many associations, they were cached as an array (with
a single object in it).  This commit switches that to use the
same type of caching at many_to_one associations, were nil is
a cached negative lookup, and where a non-nil value is the cached
lookup.

While here, add a limit(1) to the many_to_one association (in two
places, one of which will be removed soon).

There is a slight breakage of backwards compatibility in regards
to *_to_one associations with a nil :key option still addining a
limit(1), but that will be fixed soon.