Raise an error if attempting to create a one_to_many association with a :one_to_one option
Because backwards compatibility has been broken significantly for
one_to_one associations, I feel it is best to have Sequel raise an
error right away when attempting to define the association, rather
that proceed and raise an error later (e.g. when eager loading).

This also makes the code work if associate is called.  Since
associate is a public method, users are allowed to call it directly,
and that case should be handled.

one_to_one associations no longer use the :one_to_one option, since
we can just check the :type option instead.

This should be the final one_to_one related commit for the time
being.