Add the :eager_loader_key option to associations, useful for custom eager loaders
Advanced custom eager loaders may want to use a key for the eager
loader that isn't the same as the common :key option.  This can
be especially important in that if the eager loader key is empty,
there is no attempt made to eager load the association (since there
would be no matching records returned).