Add Sequel::BasicObject.remove_methods!, useful on 1.8 if libraries required after Sequel add methods to Object
This also removes private methods from Sequel::BasicObject,
previously only public and protected methods were removed.

I first realized this is a problem when coming up with some example
code using bin/sequel to help a user:

  DB[:a].filter{x > y}

This actually raised an error, because YAML adds the private Kernel
method y, and that wasn't removed.  Originally, I thought the reason
was that YAML was required after Sequel, but it turns out that it
was required before Sequel (since rubygems requires it), and the
problem was that Sequel::BasicObject did not remove any private
methods, which I consider to be a bug.

Since it's possible to require libraries after Sequel which may
add methods to Object that you want to remove,
Sequel::BasicObject.remove_methods! exists to make sure that those
methods will not override the method_missing handling in
Sequel::BasicObject subclasses.

On ruby 1.9, Sequel::BasicObject.remove_methods! exists for
compatibility, but it's a noop.