Turn DISTINCT ON support off by default, and emulate it with GROUP BY on MySQL
DISTINCT ON is a PostgreSQL extension, but Sequel had it being turned
on by default, with MySQL, Oracle, and SQLite turning it off.
However, as far as I know, no other database supports it, so it
really should be off by default.

While MySQL does not support DISTINCT ON, it can emulate PostgreSQL's
support to a certain extent using it's non-standard GROUP BY
implementation, so Sequel will use GROUP BY instead of DISTINCT ON
on MySQL.  The main difference is that on PostgreSQL, DISTINCT ON
can be made deterministic through ordering, which isn't the case
on MySQL.

This adds adapter specs for this support for both MySQL and
PostgreSQL.