Add support for :test option to Database to be automatically test the connection
The default setting of bin/sequel has been to test the connection
before bringing up an IRB shell, but the previously Sequel had no
way of doing this automatically (bin/sequel just called
test_connection manually after instantiating the database object).

This commit allows Sequel.connect and related methods to take a
:test option that specifies whether a connection attempt should
be made directly after connecting.  Like other connection options,
on adapters other than do and jdbc, you can specify the option in
the connection string:

  Sequel.connect('postgres://host/database?test=t')

It is possibly that this will be the default in a later version of
Sequel, though I do not have plans to do that currently.

While here, switch bin/sequel over to using this new capability.