Add foreign_keys accessor to SQLite Database objects, which modifies the foreign_keys PRAGMA available in 3.6.19+
I hadn't noticed this before, but all PRAGMA modification operations
in the shared SQLite adapter are not thread safe.  The only reason
this wasn't caught before is because it's likely few people use them
and few people run SQLite in a threaded environment, and the
intersection of those groups is probably very small.

A future commit will address this by making all PRAGMA modification
operators apply directly on the underlying connection objects
inside Database#connect, by specifying the options as database
options when calling Sequel.connect.  That's the only thread-safe
simple way to do it.